requirements.txt
python-telegram-bot==21.7  # Последняя стабильная версия, совместимая с Python 3.13
python-dotenv==1.0.0
jsonschema==4.20.0